+++ src/runCheck.ts
@@ -7,9 +7,11 @@ import { Merger } from "./workflow/merger";
 import { WorkflowResource } from "./workflow/resource";
 import { Workflow } from "./workflow/workflow";
 
-type TCheckStatus = "error" | "hasupdates" | "updated";
-
-function logCheck(check: Check, forceUpdate: boolean, showDiff: boolean) {
+export function logCheck(
+    check: Check,
+    forceUpdate = false,
+    showDiff = false
+): void {
     if (check.action === "equal") return;
     if (!forceUpdate && check.force) {
         return console.log(chalk.grey(`  ${check.noForceMessage}`));
@@ -18,73 +20,56 @@ function logCheck(check: Check, forceUpdate: boolean, showDiff: boolean) {
     if (showDiff) logDiff(check.oldValue, check.newValue);
 }
 
-export function runCheck(
-    workflowItem: WorkflowResource,
-    localContent: string,
-    remoteContent: string,
-    forceUpdate: boolean,
-    showDiff: boolean
-): TCheckStatus {
-    let localWorkflow: Workflow;
-    try {
-        localWorkflow = Workflow.fromString(localContent);
-    } catch (e) {
-        console.log(chalk.red(`  ✗  ${e}`));
-        return "error";
-    }
-    const remoteWorkflow = Workflow.fromString(remoteContent);
-    remoteWorkflow.commentLines = workflowItem.getCommentLines();
+export function getCheckResult(
+    resource: WorkflowResource,
+    checks: Array<Check>,
+    forceUpdate: boolean
+): Check {
+    const errorChecks = checks.filter(check => check.isError());
+    if (errorChecks.length) return new Check("cancelled", "error");
 
+    if (!resource.existsLocally()) return new Check("workflow", "added");
+    const applyChecks = checks.filter(check => check.isApplied(forceUpdate));
+    if (applyChecks.length) return new Check("workflow", "updated");
+
+    return new Check("workflow", "up to date");
+}
+
+export async function runCheck(
+    workflowItem: WorkflowResource,
+    forceUpdate: boolean
+): Promise<Array<Check>> {
+    if (!workflowItem.existsLocally()) {
+        return [];
+    }
+    const localWorkflow = await workflowItem.getLocal();
+    let remoteWorkflow: Workflow;
+    try {
+        remoteWorkflow = await workflowItem.getRemote();
+    } catch (e) {
+        return [new Check(`${e}`, "error")];
+    }
     const checker = new Checker(forceUpdate, localWorkflow);
 
-    const errors = checker.getErrors();
-    if (errors.length) {
-        errors.forEach(error => console.log(chalk.red(`  ✗  ${error}`)));
-        console.log(
-            chalk.red(`  ✗  has ${chalk.bold("errors")} that prevent update`)
-        );
-        return "error";
-    }
     const newWorkflow = new Merger(true).merge(localWorkflow, remoteWorkflow);
-    const checks = checker.getChecks(newWorkflow);
-    const applyChecks = checks.filter(check => check.isApplied(forceUpdate));
-
-    checks.forEach(check => logCheck(check, forceUpdate, showDiff));
-    if (!applyChecks.length) {
-        console.log(chalk.grey("  ✓  is up to date"));
-        return "updated";
-    }
-    console.log(chalk.green("  ✓  can be updated"));
-    return "hasupdates";
+    return checker.getChecks(newWorkflow);
 }
 
 export async function runCheckAll(
-    items: Array<WorkflowResource>,
+    resources: Array<WorkflowResource>,
     forceUpdate: boolean,
     showDiff: boolean
 ): Promise<boolean> {
-    const results: Array<Promise<TCheckStatus>> = items.map(async item => {
-        const remoteContent = await item.getRemote();
-        const localContent = await item.getLocal();
-        console.log(item.getTitle());
-        if (!remoteContent) {
-            console.log(chalk.red(`  ✗  download failed: ${item.url}`));
-            return "error";
-        }
-        if (!localContent) {
-            console.log(chalk.green("  ✓  will be created"));
-            return "hasupdates";
-        }
-
-        return runCheck(
-            item,
-            localContent,
-            remoteContent,
-            forceUpdate,
-            showDiff
-        );
-    });
-    return Promise.all(results).then(
-        results => results.filter(x => x === "error").length === 0
+    const checkLists = await Promise.all(
+        resources.map(resource => runCheck(resource, forceUpdate))
     );
+    const errorChecks: Array<Check> = [];
+    resources.forEach((resource, index) => {
+        console.log(resource.getTitle());
+        const checks = checkLists[index];
+        errorChecks.push(...checks.filter(check => check.isError()));
+        checks.forEach(check => logCheck(check, forceUpdate, showDiff));
+        logCheck(getCheckResult(resource, checks, forceUpdate), false, false);
+    });
+    return errorChecks.length === 0;
 }